promotion updated

+++ database/migrations/2024_07_28_121234_create_promotions_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->decimal('discount_amount', 8, 2);
             $table->decimal('original_price', 8, 2);
             $table->string('type');
-            $table->decimal('limit', 8, 2);
+            $table->integer('limit');
             $table->date('start_at');
             $table->date('end_at');
             $table->string('status');
